refactor(templates): extract the script-template library into ferrowl-templates
Move the bundled Lua templates, the build.rs codegen, and the ScriptTemplate
type into a new ferrowl-templates crate (13th workspace member, lockstep
versioned). The crate carries its own TemplateContext; the binary keeps
ScriptContext and maps at the boundary via From<TemplateContext>, owning the
ScriptContext-keyed templates() filter in script_template.rs. OCPP template
dirs regrouped under templates/ocpp/{client,server}. Same compiled-in set and
metadata as before (SC-R-036/037); their tests move with the code (SC-R-037 to
ferrowl-templates, SC-R-036 filtering stays in ferrowl).
