+++ ferrowl-templates/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "ferrowl-templates"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+publish.workspace = true
+
+[lints]
+workspace = true
+
+[dev-dependencies]
+ferrowl-lua = { path = "../ferrowl-lua" }
