+++ ferrowl-templates/templates/modbus/apply-limit.lua
