+++ ferrowl-templates/templates/modbus/register-mirror.lua
