+++ ferrowl-templates/templates/modbus/register-ramp.lua
