+++ ferrowl-templates/templates/modbus/sine-wave.lua
