+++ ferrowl-templates/templates/ocpp/client/meter-values.lua
