+++ ferrowl-templates/templates/session/module-inventory.lua
