+++ ferrowl-templates/templates/session/power-report.lua
